+++ frontend/src/components/common/Tabs/TabList.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { css, styled } from 'styled-components';
 
 import color from '@Styles/color';
@@ -7,76 +7,13 @@ import LeftArrow from '@Assets/icons/LeftArrowIcon';
 import RightArrow from '@Assets/icons/RightArrowIcon';
 
 import { useTabs } from './TabsContext';
+import useTabListScroll from './hooks/useTabListScroll';
 
 const TabList = () => {
-  const { tabs, selectedTab, changeTab } = useTabs();
-
-  const [isOverFlow, setIsOverFlow] = useState(false);
-  const [scrollButtonPosition, setScrollButtonPosition] = useState<'start' | 'end' | 'both' | null>(null);
-
-  const hasStartScrollButton = scrollButtonPosition === 'start' || scrollButtonPosition === 'both';
-  const hasEndScrollButton = scrollButtonPosition === 'end' || scrollButtonPosition === 'both';
-
   const tabList = useRef<HTMLUListElement>(null);
 
-  const handleMoveScroll = (position: 'start' | 'end') => {
-    if (!tabList.current) return;
-
-    const { scrollLeft, clientWidth } = tabList.current;
-
-    const movedPosition = scrollLeft + (position === 'end' ? +(clientWidth / 2) : -(clientWidth / 2));
-
-    tabList.current.scrollTo({
-      left: movedPosition,
-      behavior: 'smooth',
-    });
-  };
-
-  const handleScroll = useCallback(() => {
-    const { scrollWidth, clientWidth, scrollLeft } = tabList.current!;
-
-    // scrollWidth: 스크롤까지 포함한 전체 너비
-    // clientWidth: 실제 화면에 보이는 너비
-    // scrollLeft: 가로로 스크롤한 너비
-    const isEndOfList = Math.abs(scrollWidth - scrollLeft - clientWidth) < 1;
-
-    // 스크롤이 끝 지점으로 이동했을 때
-    if (isEndOfList) setScrollButtonPosition('start');
-    // 스크롤이 앞 지점에 있을 때
-    else if (scrollLeft === 0) setScrollButtonPosition('end');
-    // 양쪽에 스크롤이 생길 때
-    else setScrollButtonPosition('both');
-
-    if (!isOverFlow) setScrollButtonPosition(null);
-  }, [isOverFlow]);
-
-  const handleTabListResize: ResizeObserverCallback = useCallback(
-    ([tabList]) => {
-      const { scrollWidth, clientWidth } = tabList.target;
-
-      if (scrollWidth > clientWidth) setIsOverFlow(true);
-      else setIsOverFlow(false);
-
-      handleScroll();
-    },
-    [handleScroll],
-  );
-
-  useEffect(() => {
-    const tabListResizeObserver = new ResizeObserver(handleTabListResize);
-
-    if (!tabList.current) return;
-
-    tabListResizeObserver.observe(tabList.current);
-    tabList.current.addEventListener('scroll', handleScroll);
-
-    return () => {
-      if (!tabList.current) return;
-
-      tabListResizeObserver.unobserve(tabList.current);
-      tabList.current.removeEventListener('scroll', handleScroll);
-    };
-  }, [handleScroll, handleTabListResize]);
+  const { tabs, selectedTab, changeTab } = useTabs();
+  const { hasStartScrollButton, hasEndScrollButton, handleMoveScroll } = useTabListScroll(tabList);
 
   return (
     <Layout>
